refactor: SharedDesignSystem에 있던 임시 TXItem 구현체 Feature에 맞게 분리 - #278
